The Drawobject is painting a rabbit to the right.. =)

+++ drawable.h
@@ -3,8 +3,9 @@
 #include "VectorUtils3.h"
 #include "GL_utilities.h"
 #include "controller.h"
+#include "constants.h"
 
-class DrawableObject{
+class DrawableObject {
 public:
 	mat4 trans;
 	mat4 rot;
@@ -13,7 +14,7 @@ public:
 	
 	DrawableObject();
 
-	DrawableObject(int x, int yOffset, int z, GLfloat rotation, GLuint* tex, Model* model, GLuint* program);
+	DrawableObject(GLfloat x, GLfloat yOffset, GLfloat z, GLfloat rotation, GLuint* tex, Model* model, GLuint* program);
 
 	DrawableObject(vec3 position, GLfloat rotation, GLuint* tex, Model* model, GLuint* program);
 
@@ -26,18 +27,18 @@ public:
 	void rotate(GLfloat angle);
 	
 	//use 0,1,0 to move only along the y-axis. See also: setCoords()
-	void move(int x, int y, int z);
+	void move(GLfloat x, GLfloat y, GLfloat z);
 
 	//SETS rotation (rotation = angle). See also: rotate(GLfloat)
 	void setRotation(GLfloat angle);
 
 	//use NULL,0,NULL to set y = 0 while not affecting x or z. See also move()
-	void setCoords(int x, int y, int z);
+	void setCoords(GLfloat x, GLfloat y, GLfloat z);
 	vec3 getCoords();
 
 private:
 	void updateMatrices();
-	int x, z, y, yOffset; //yOffset is distance from ground
+	GLfloat x, z, y, yOffset; //yOffset is distance from ground
 	GLfloat rotation;
 	Model* model;
 	GLuint* program;
